requirements.txt
streamlit
gitpython
docker
google-generativeai
pyyaml
pytest
